requirements.txt
argcomplete>=3.1,<4
pillow
psutil
wcwidth

# AI (aienhance) dependencies are installed via install.sh when enabled:
# numpy<2
# opencv-python-headless<4.12
